Refactor Gen3 customization code into subpackage.
